+++ models/TabUserModel.tsx
@@ -1,16 +1,16 @@
 import React, { useEffect, useState, useRef } from 'react';
-import './stylesUserProfile.css';
+import '../styles/stylesUserProfile.css';
 import { RiContactsLine } from 'react-icons/ri';
 import { CiStar } from 'react-icons/ci';
-import { getAuthToken } from './background';
 import SubscriptionModel from './SubscriptionModel';
+import { getAuthToken } from '../background';
 interface ProfileInfo {
   names?: { displayName: string }[];
   emailAddresses?: { value: string }[];
   photos?: { url: string }[];
 }
 
-const TabUserProfile: React.FC = () => {
+const TabUserModel: React.FC = () => {
   const [responseText, setResponseText] = useState<ProfileInfo | null>(null);
   const [loading, setLoading] = useState<boolean>(true);
   const [activeModule, setActiveModule] = useState<string>('Profile');
@@ -207,4 +207,4 @@ const TabUserProfile: React.FC = () => {
   );
 };
 
-export default TabUserProfile;
+export default TabUserModel;
