+++ public/auth.html
