+++ public/auth.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import AuthModel from './AuthModel';
+import AuthModel from '../models/AuthModel';
 const root = createRoot(document.getElementById('auth') as HTMLElement);
 root.render(<AuthModel />);
