+++ public/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from '../App';
 const root = createRoot(document.getElementById('root') as HTMLElement);
 root.render(<App />);
