+++ public/infoModel.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import UserProfile from './UserProfile';
+import UserModel from '../models/UserModel';
 const root = createRoot(document.getElementById('infoModel') as HTMLElement);
-root.render(<UserProfile />);
+root.render(<UserModel />);
