+++ public/tabInfoModel.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import TabUserProfile from './TabUserProfile';
+import TabUserModel from '../models/TabUserModel';
 const root = createRoot(document.getElementById('tabInfoModel') as HTMLElement);
-root.render(<TabUserProfile />);
+root.render(<TabUserModel />);
