--- subscription.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <div id="subscription"></div>
-    <script src="./subscription.js"></script>
-  </body>
-</html>
